--- content_export_and_import.links.task.yml
@@ -1,5 +1,0 @@
-content_export_and_import.export_node:
-  title: 'Export'
-  route_name: content_export_and_import.export_node
-  base_route: entity.node.canonical
-  weight: 100
