--- content_export_and_import.module
@@ -1,37 +1,0 @@
-<?php
-
-/**
- * @file
- * Contains content_export_and_import.module.
- */
-
-use Drupal\Core\Routing\RouteMatchInterface;
-
-/**
- * Implements hook_help().
- */
-function content_export_and_import_help($route_name, RouteMatchInterface $route_match) {
-  switch ($route_name) {
-    // Main module help for the content_export_and_import module.
-    case 'help.page.content_export_and_import':
-      $output = '';
-      $output .= '<h3>' . t('About') . '</h3>';
-      $output .= '<p>' . t('Content export and import') . '</p>';
-      return $output;
-
-    default:
-  }
-}
-
-/**
- * Implements hook_theme().
- */
-function content_export_and_import_theme($existing, $type, $theme, $path) {
-  return [
-    'content_export_and_import_index' => [
-      'variables' => [
-        'exports' => NULL,
-      ],
-    ],
-  ];
-}
