--- content_export_and_import.routing.yml
@@ -1,35 +1,0 @@
-content_export_and_import.index:
-  path: '/content_export_and_import'
-  defaults:
-    _controller: '\Drupal\content_export_and_import\Controller\ExportController::index'
-    _title: 'Content exports'
-  requirements:
-    # @todo
-    _permission: 'access content'
-    # _custom_access: '\Drupal\content_export_and_import\Controller\ExportController::access'
-
-content_export_and_import.import:
-  path: '/content_export_and_import/import/{entity_type}/{entity_id}'
-  defaults:
-    _controller: '\Drupal\content_export_and_import\Controller\ExportController::import'
-    _title_callback: '\Drupal\content_export_and_import\Controller\ExportController::importTitle'
-  requirements:
-    # @todo
-    _permission: 'access content'
-    _custom_access: '\Drupal\content_export_and_import\Controller\ExportController::importAccess'
-
-content_export_and_import.export:
-  path: '/content_export_and_import/export/{entity_type}/{entity_id}'
-  defaults:
-    _controller: '\Drupal\content_export_and_import\Controller\ExportController::export'
-    _title_callback: '\Drupal\content_export_and_import\Controller\ExportController::exportTitle'
-  requirements:
-    _custom_access: '\Drupal\content_export_and_import\Controller\ExportController::exportAccess'
-
-content_export_and_import.export_node:
-  path: '/content_export_and_import/export/node/{node}'
-  defaults:
-    _controller: '\Drupal\content_export_and_import\Controller\ExportController::exportNode'
-    _title_callback: '\Drupal\content_export_and_import\Controller\ExportController::exportNodeTitle'
-  requirements:
-    _custom_access: '\Drupal\content_export_and_import\Controller\ExportController::exportNodeAccess'
