--- content_export_and_import.services.yml
@@ -1,11 +1,0 @@
-services:
-  content_export_and_import.entity_exporter:
-    class: Drupal\content_export_and_import\Export\EntityExporter
-    lazy: true
-    arguments:
-      - '@serializer'
-      - '@file_system'
-
-  content_export_and_import.entity_importer:
-    class: Drupal\content_export_and_import\Export\EntityImporter
-    lazy: true
