+++ entity_export_and_import.links.task.yml
@@ -1,0 +1,5 @@
+entity_export_and_import.export_node:
+  title: 'Export'
+  route_name: entity_export_and_import.export_node
+  base_route: entity.node.canonical
+  weight: 100
