+++ entity_export_and_import.module
@@ -1,0 +1,37 @@
+<?php
+
+/**
+ * @file
+ * Contains entity_export_and_import.module.
+ */
+
+use Drupal\Core\Routing\RouteMatchInterface;
+
+/**
+ * Implements hook_help().
+ */
+function entity_export_and_import_help($route_name, RouteMatchInterface $route_match) {
+  switch ($route_name) {
+    // Main module help for the entity_export_and_import module.
+    case 'help.page.entity_export_and_import':
+      $output = '';
+      $output .= '<h3>' . t('About') . '</h3>';
+      $output .= '<p>' . t('Entity export and import') . '</p>';
+      return $output;
+
+    default:
+  }
+}
+
+/**
+ * Implements hook_theme().
+ */
+function entity_export_and_import_theme($existing, $type, $theme, $path) {
+  return [
+    'entity_export_and_import_index' => [
+      'variables' => [
+        'exports' => NULL,
+      ],
+    ],
+  ];
+}
