+++ src/EntityImporter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\content_export_and_import\Export;
+namespace Drupal\entity_export_and_import;
 
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Entity\EntityStorageException;
@@ -13,7 +13,7 @@ class EntityImporter implements EntityImporterInterface {
   /**
    * {@inheritdoc}
    */
-  public function importEntity(EntityInterface $entity) {
+  public function importEntity(EntityInterface $entity): bool {
     try {
       $entity
         ->enforceIsNew()
