+++ src/Form/ExportForm.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Drupal\content_export_and_import\Form;
+namespace Drupal\entity_export_and_import\Form;
 
-use Drupal\content_export_and_import\Export\EntityExporterInterface;
+use Drupal\entity_export_and_import\EntityExporterInterface;
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
@@ -15,7 +15,7 @@ class ExportForm extends FormBase {
   /**
    * The entity exporter.
    *
-   * @var \Drupal\content_export_and_import\Export\EntityExporterInterface
+   * @var \Drupal\entity_export_and_import\EntityExporterInterface
    */
   private $entityExporter;
 
@@ -38,7 +38,7 @@ class ExportForm extends FormBase {
    */
   public static function create(ContainerInterface $container) {
     return new static(
-      $container->get('content_export_and_import.entity_exporter')
+      $container->get('entity_export_and_import.entity_exporter')
     );
   }
 
@@ -55,7 +55,7 @@ class ExportForm extends FormBase {
    * {@inheritdoc}
    */
   public function getFormId() {
-    return 'content_export_and_import_export';
+    return 'entity_export_and_import_export';
   }
 
   /**
@@ -75,7 +75,7 @@ class ExportForm extends FormBase {
       '#button_type' => 'primary',
     ];
 
-    $form['#attached']['library'][] = 'content_export_and_import/admin';
+    $form['#attached']['library'][] = 'entity_export_and_import/admin';
 
     return $form;
   }
