+++ templates/entity-export-and-import-index.html.twig
@@ -1,0 +1,13 @@
+<div class="entity-export-and-import--exports">
+  {% if exports %}
+    <ul>
+      {% for export in exports %}
+        <li>
+          <a href="{{ path('entity_export_and_import.import', export) }}">{{ export.entity_type }}/{{ export.entity_id }}</a>
+        </li>
+      {% endfor %}
+    </ul>
+  {% else %}
+    <p>{{ 'No entity exports found'|t }}</p>
+  {% endif %}
+</div>
